seperate out NRT tests
